Project_1-TicTacToe
Commit and push changes done by GitHub Desktop application:

1. The basic form of the tic-tac-toe game.
2. First modification: added the ability to define the board size.
3. Second modification: added the ability to start a new game or end the game.
4. Third modification: added requirement to enter players' names/nicknames before starting the game, and a summary after the game, including statistics.

I think that "in the future" I could try to optimize the code using lambda expressions for collections, loops and if-else conditional statements.

Moreover, I think that displaying the game summary could also be solved using an enum and an enum loop (useful in games with a large number of players at the same time, and in our case there are only X and 0).